requirements.txt
selenium==4.1.3
pytest==7.1.1
webdriver-manager==3.5.4
allure-pytest==2.9.45
pytest-xdist==2.5.0
flake8==4.0.1
black==22.3.0
mysql-connector-python==8.0.23
psycopg2-binary==2.8.6
sqlalchemy==1.4.28
